Crate oci_spec

Structs

Box
Box specifies dimensions of a rectangle. Used for specifying the size of a console.

Hook
Hook specifies a command that is run at a particular event in the lifecycle of a container.

Hooks
Hooks specifies a command that is run in the container at a particular event in the lifecycle (setup and teardown) of a container.

Linux
Linux contains platform-specific configuration for Linux based containers.

LinuxBlockIo
LinuxBlockIO for Linux cgroup ‘blkio’ resource management.

LinuxCapabilities
LinuxCapabilities specifies the list of allowed capabilities that are kept for a process. http://man7.org/linux/man-pages/man7/capabilities.7.html

LinuxCpu
LinuxCPU for Linux cgroup ‘cpu’ resource management.

LinuxDevice
LinuxDevice represents the mknod information for a Linux special device file.

LinuxDeviceCgroup
Represents a device rule for the devices specified to the device controller

LinuxHugepageLimit
LinuxHugepageLimit structure corresponds to limiting kernel hugepages.

LinuxIdMapping
LinuxIDMapping specifies UID/GID mappings.

LinuxIntelRdt
LinuxIntelRdt has container runtime resource constraints for Intel RDT CAT and MBA features which introduced in Linux 4.10 and 4.12 kernel.

LinuxInterfacePriority
LinuxInterfacePriority for network interfaces.

LinuxMemory
LinuxMemory for Linux cgroup ‘memory’ resource management.

LinuxNamespace
LinuxNamespace is the configuration for a Linux namespace.

LinuxNetwork
LinuxNetwork identification and priority configuration.

LinuxPersonality
LinuxPersonality represents the Linux personality syscall input.

LinuxPids
LinuxPids for Linux cgroup ‘pids’ resource management (Linux 4.3).

LinuxRdma
LinuxRdma for Linux cgroup ‘rdma’ resource management (Linux 4.11).

LinuxResources
Resource constraints for container

LinuxRlimit
RLimit types and restrictions.

LinuxSeccomp
LinuxSeccomp represents syscall restrictions.

LinuxSeccompArg
LinuxSeccompArg used for matching specific syscall arguments in seccomp.

LinuxSyscall
LinuxSyscall is used to match a syscall in seccomp.

LinuxThrottleDevice
LinuxThrottleDevice struct holds a major:minor rate_per_second pair.

LinuxWeightDevice
LinuxWeightDevice struct holds a major:minor weight pair for weightDevice.

Mount
Mount specifies a mount for a container.

Process
Process contains information to start a specific application inside the container.

Root
Root contains information about the container’s root filesystem on the host.

Solaris
Solaris contains platform-specific configuration for Solaris application containers.

SolarisAnet
SolarisAnet provides the specification for automatic creation of network resources for this container.

SolarisCappedCPU
SolarisCappedCPU allows users to set limit on the amount of CPU time that can be used by container.

SolarisCappedMemory
SolarisCappedMemory allows users to set the physical and swap caps on the memory that can be used by this container.

Spec
Base configuration for the container.

User
User id (uid) and group id (gid) tracks file permssions.

VM
VM contains information for virtual-machine-based containers.

VMHypervisor
VMHypervisor contains information about the hypervisor to use for a virtual machine.

VMImage
VMImage contains information about the virtual machine root image.

VMKernel
VMKernel contains information about the kernel to use for a virtual machine.

Windows
Windows defines the runtime configuration for Windows based containers, including Hyper-V containers.

WindowsCPUResources
WindowsCPUResources contains CPU resource management settings.

WindowsDevice
WindowsDevice represents information about a host device to be mapped into the container.

WindowsHyperV
WindowsHyperV contains information for configuring a container to run with Hyper-V isolation.

WindowsMemoryResources
WindowsMemoryResources contains memory resource management settings.

WindowsNetwork
WindowsNetwork contains network settings for Windows containers.

WindowsResources

WindowsStorageResources
WindowsStorageResources contains storage resource management settings.

Enums

Arch

FreezerState

LinuxDeviceType
Device types

LinuxNamespaceType

LinuxPersonalityDomain
Define domain and flags for LinuxPersonality.

LinuxRlimitType
Available rlimti types (see https://man7.org/linux/man-pages/man2/getrlimit.2.html)

LinuxSeccompAction

LinuxSeccompOperator

Functions

get_default_maskedpaths

get_default_mounts

get_default_namespaces

get_default_readonly_paths