oci_spec
Module distribution

OCI distribution spec types and definitions.

The Open Container Initiative Distribution Specification (a.k.a. “OCI Distribution Spec”) defines an API protocol to facilitate and standardize the distribution of content.

While OCI Image is the most prominent, the specification is designed to be agnostic of content types. Concepts such as “manifests” and “digests”, are currently defined in the Open Container Initiative Image Format Specification (a.k.a. “OCI Image Spec”).

To support other artifact types, please see the Open Container Initiative Artifact Authors Guide (a.k.a. “OCI Artifacts”).

Structs

ErrorInfo

Describes a server error returned from a registry.

ErrorInfoBuilder

Builder for ErrorInfo.

ErrorResponse

ErrorResponse is returned by a registry on an invalid request.

ErrorResponseBuilder

Builder for ErrorResponse.

Reference

Reference provides a general type to represent any way of referencing images within an OCI registry.

RepositoryList

RepositoryList returns a catalog of repositories maintained on the registry.

RepositoryListBuilder

Builder for RepositoryList.

TagList

A list of tags for a given repository.

TagListBuilder

Builder for TagList.

Enums

ErrorCode

Unique identifier representing error code.

ParseError

Reasons that parsing a string as a Reference can fail.

Constants

ERR_REGISTRY

The string returned by and ErrorResponse error.

VERSION_DEV

Indicates development branch. Releases will be empty string.

VERSION_MAJOR

API incompatible changes.

VERSION_MINOR

Changing functionality in a backwards-compatible manner

VERSION_PATCH

Backwards-compatible bug fixes.

Functions

version

Retrieve the version as string representation.