Enum oci_spec::image::MediaType

pub enum MediaType {

    Descriptor,
    LayoutHeader,
    ImageManifest,
    ImageIndex,
    ImageLayer,
    ImageLayerGzip,
    ImageLayerZstd,
    ImageLayerNonDistributable,
    ImageLayerNonDistributableGzip,
    ImageLayerNonDistributableZstd,
    ImageConfig,
    ArtifactManifest,
    Other(String),

}

Media types used by OCI image format spec. Values MUST comply with RFC 6838, including the naming requirements in its section 4.2.

Variants

Descriptor

MediaType Descriptor specifies the media type for a content descriptor.

LayoutHeader

MediaType LayoutHeader specifies the media type for the oci-layout.

ImageManifest

MediaType ImageManifest specifies the media type for an image manifest.

ImageIndex

MediaType ImageIndex specifies the media type for an image index.

ImageLayer

MediaType ImageLayer is the media type used for layers referenced by the manifest.

ImageLayerGzip

MediaType ImageLayerGzip is the media type used for gzipped layers referenced by the manifest.

ImageLayerZstd

MediaType ImageLayerZstd is the media type used for zstd compressed layers referenced by the manifest.

ImageLayerNonDistributable

MediaType ImageLayerNonDistributable is the media type for layers referenced by the manifest but with distribution restrictions.

ImageLayerNonDistributableGzip

MediaType ImageLayerNonDistributableGzip is the media type for gzipped layers referenced by the manifest but with distribution restrictions.

ImageLayerNonDistributableZstd

MediaType ImageLayerNonDistributableZstd is the media type for zstd compressed layers referenced by the manifest but with distribution restrictions.

ImageConfig

MediaType ImageConfig specifies the media type for the image configuration.

ArtifactManifest

MediaType ArtifactManifest specifies the media type used for content addressable artifacts to store them along side container images in a registry.

Other(String)

MediaType not specified by OCI image format.