Struct oci_spec::image::ConfigBuilder

pub struct ConfigBuilder { /* private fields */ }

Builder for Config.

Implementations

impl ConfigBuilder

pub fn user<VALUE: Into<String>>(self, value: VALUE) -> Self

The username or UID which is a platform-specific structure that allows specific control over which user the process run as. This acts as a default value to use when the value is not specified when creating a container. For Linux based systems, all of the following are valid: user, uid, user:group, uid:gid, uid:group, user:gid. If group/gid is not specified, the default group and supplementary groups of the given user/uid in /etc/passwd from the container are applied.

pub fn exposed_ports<VALUE: Into<Vec<String>>>(self, value: VALUE) -> Self

A set of ports to expose from a container running this image. Its keys can be in the format of: port/tcp, port/udp, port with the default protocol being tcp if not specified. These values act as defaults and are merged with any specified when creating a container.

pub fn env<VALUE: Into<Vec<String>>>(self, value: VALUE) -> Self

Entries are in the format of VARNAME=VARVALUE. These values act as defaults and are merged with any specified when creating a container.

pub fn entrypoint<VALUE: Into<Vec<String>>>(self, value: VALUE) -> Self

A list of arguments to use as the command to execute when the container starts. These values act as defaults and may be replaced by an entrypoint specified when creating a container.

pub fn cmd<VALUE: Into<Vec<String>>>(self, value: VALUE) -> Self

Default arguments to the entrypoint of the container. These values act as defaults and may be replaced by any specified when creating a container. If an Entrypoint value is not specified, then the first entry of the Cmd array SHOULD be interpreted as the executable to run.

pub fn volumes<VALUE: Into<Vec<String>>>(self, value: VALUE) -> Self

A set of directories describing where the process is likely to write data specific to a container instance.

pub fn working_dir<VALUE: Into<String>>(self, value: VALUE) -> Self

Sets the current working directory of the entrypoint process in the container. This value acts as a default and may be replaced by a working directory specified when creating a container.

pub fn labels<VALUE: Into<HashMap<String, String>>>(self, value: VALUE) -> Self

The field contains arbitrary metadata for the container. This property MUST use the annotation rules.

pub fn stop_signal<VALUE: Into<String>>(self, value: VALUE) -> Self

The field contains the system call signal that will be sent to the container to exit. The signal can be a signal name in the format SIGNAME, for instance SIGKILL or SIGRTMIN+3.

pub fn build(self) -> Result<Config, OciSpecError>

Builds a new Config.

Errors

If a required field has not been initialized.